Break out table rendering from `pkg/cmd/pulumi/util`
This commit continues the process of breaking up
`pkg/cmd/pulumi/util.go` by hoisting out functions for printing tabular
data to a new `ui` subpackage.
